+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-# 1.8.2
-- Fix UniteStonesOfAll
+# 1.8.3
+- MineCarve's effect now has duration ticks
+- Final Power now ignores enchantments and potion effects